Moved somethings around. Added time to tasks
