+++ src/components/Header.jsx
@@ -1,17 +1,17 @@
 import "./Header.css";
-import { useAppData } from "../themes/AppTheme";
-import { useTask } from "../themes/TaskTheme";
+import { useAppData } from "../contexts/AppContext";
+import { useTask } from "../contexts/TaskContext";
 
 export const Header = () => {
-    const { name, appContent } = useAppData();
-    const { taskData, addTask } = useTask();
+  const { name, appContent } = useAppData();
+  const { taskData, addTask } = useTask();
 
   console.log(name, appContent);
 
   return (
     <header>
-          <h1>{appContent.heading}</h1>
-          <p>Task count: {taskData.length}</p>
+      <h1>{appContent.heading}</h1>
+      <p>Task count: {taskData ? taskData.length : "0"}</p>
     </header>
   );
 };
